Use object transformer
