--- internal/export/transform_context.go
@@ -1,9 +1,0 @@
-// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-// or more contributor license agreements. Licensed under the Elastic License;
-// you may not use this file except in compliance with the Elastic License.
-
-package export
-
-type transformationContext struct {
-	packageName string
-}
